# swagger.yaml
openapi: 3.0.3
info:
  title: DOK Live API
  description: |
    Complete REST API documentation for the DOK Live educational platform.
    
    ## Authentication
    Most endpoints require a JWT Bearer token obtained from `POST /login`.
    - **Admin token** — returned when an admin/assistant logs in
    - **Student token** — returned when a student logs in
    
    Pass the token as: `Authorization: Bearer <token>`
  version: 1.0.0
  contact:
    name: DOK Live
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://dok-edu.com
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: Error
        data:
          type: object
          properties:
            message:
              type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: Error
            data:
              message: Not authorized, no token
...
paths:
...
  /login:
    post:
      tags: [Auth]
      summary: Log in (admin or student)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  example: secret123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Login successful
                  token:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                      email:
                        type: string
                      group:
                        type: string
                      name:
                        type: string
                      role:
                        type: string
                        example: teacher
        '400':
          description: Wrong password (admin)
          content:
            application/json:
              example:
                status: Error
                data:
                  message: Invalid email or password
        '401':
          description: Wrong password (student)
          content:
            application/json:
              example:
                status: Error
                data:
                  message: Wrong password
        '403':
          description: Account not verified or banned
          content:
            application/json:
              examples:
                notVerified:
                  value:
                    status: Error
                    data:
                      message: Email not verified
                banned:
                  value:
                    status: Error
                    data:
                      message: Your account is banned
        '404':
          description: Email not found
          content:
            application/json:
              example:
                status: Error
                data:
                  message: Email not found
...
  /feed:
    get:
      tags: [Feed]
      summary: Get all feed posts (auto-purges posts older than 14 days)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Feed posts
          content:
            application/json:
              example:
                status: success
                results: 4
                data:
                  - feedId: 1
                    text: Session moved to Saturday
                    dateAndTime: '2026-08-01T08:00:00.000Z'
                    semester: jun
                    adminId: 3
        '404':
          description: Feed is empty
          content:
            application/json:
              example:
                status: Error
                data:
                  message: Feed is empty
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /leaderBoard:
    get:
      tags: [Leaderboard]
      summary: Paginated leaderboard ordered by total score
      description: |
        Returns 20 students per page sorted by totalScore descending.
        Student token — response includes the student's own rank and score.
        Admin token — response omits the student object.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
          description: Page number (20 students per page)
      responses:
        '200':
          description: Leaderboard data
          content:
            application/json:
              examples:
                studentView:
                  summary: Response for a student token
                  value:
                    status: success
                    data:
                      pagination:
                        currentPage: 1
                        totalPages: 3
                        totalStudents: 55
                      student:
                        score: 85.5
                        rank: 7
                      leaderboard:
                        - studentId: 4
                          studentName: Top Student
                          totalScore: 120
                          group: groupa
                adminView:
                  summary: Response for an admin token
                  value:
                    status: success
                    data:
                      pagination:
                        currentPage: 1
                        totalPages: 3
                        totalStudents: 55
                      leaderboard:
                        - studentId: 4
                          studentName: Top Student
                          totalScore: 120
                          group: groupa
        '401':
          $ref: '#/components/responses/Unauthorized'